fix: support plain text
The piece of code that was supposed to generate a plain text fragment
was calling `buildTextElement` (which is not enough) and was giving it
the wrong arguments: `(xmlFragment, text)` instead of `(text)`.

Let's build the plain text fragment as `render-document-file.js` does
in a few places, by invoking `xmlBuilder.buildParagraph` directly,
because that works and is tested.

Maybe that's not the way it's supposed to be done, maybe we should not
use `xmlBuilder` out of `render-document-file.js`, I don't know. At
least this solution gives the expected behavior for my use case
(generation of a DOCX from an input text that might not contain any
HTML tag - which is nevertheless valid).

As a side-effect of the change, the `escape-html` dependency is now
unused, which is quite surprising.